# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/reviews": {
            "get": {
                "description": "Get a list of Reviews.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reviews"
                ],
                "summary": "Get all Reviews.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Reviews"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Creating a new Reviews.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reviews"
                ],
                "summary": "Create New Reviews.",
                "parameters": [
                    {
                        "description": "the body to create a new Reviews",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.reviewsInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Reviews"
                        }
                    }
                }
            }
        },
        "/reviews/{id}": {
            "get": {
                "description": "Get an Reviews by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reviews"
                ],
                "summary": "Get Reviews.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reviews id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Reviews"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Delete a Reviews by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reviews"
                ],
                "summary": "Delete one Reviews.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reviews id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Update Reviews by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reviews"
                ],
                "summary": "Update Reviews.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reviews id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update reviews",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.reviewsInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Reviews"
                        }
                    }
                }
            }
        },
        "/reviews/{id}/comments": {
            "get": {
                "description": "Get all Opinion by ReviewsId.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reviews"
                ],
                "summary": "Get Comments By id Reviews.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reviews id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Comments"
                            }
                        }
                    }
                }
            }
        },
        "/reviews/{id}/opinions": {
            "get": {
                "description": "Get all Opinion by ReviewsId.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reviews"
                ],
                "summary": "Get Opinions By id Reviews.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reviews id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Opinions"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.reviewsInput": {
            "type": "object",
            "properties": {
                "article": {
                    "type": "string"
                },
                "devices_id": {
                    "type": "integer"
                },
                "name_article": {
                    "type": "string"
                }
            }
        },
...
        "models.Comments": {
            "type": "object",
            "properties": {
                "comments_section": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "devicesID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "reviewsID": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "your_name": {
                    "type": "string"
                }
            }
        },
...
        "models.Opinions": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "opinion_section": {
                    "type": "string"
                },
                "reviewsID": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "your_name": {
                    "type": "string"
                }
            }
        },
        "models.Reviews": {
            "type": "object",
            "properties": {
                "article": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "devicesID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name_article": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}